Change "that demand" to "and so in need of"
More words, I know, but my mind kept reading "that demand" as referring to the closest noun (data), when really we're referring to the two disciplines. I don't think the "and so in need of" is perfect here, but I think some change like this would help prevent readers getting caught at this point in the sentence.