Change "standard" to "heavy-duty data analysis"
I think what we're trying to get at here is that Excel won't cut it any more for these disciplines, because of the characteristics of the data they're getting and because of what they're trying to do with it. If that's what we're saying, I'm not sure "standard" is the right word. Excel could be "standard", if it's what everyone in the field is using. I think the reason they're moving toward R, in the context of what we've just said, is that they need a more powerful tool.