Change "It follows that" to "Many ... also"
Here, I'm not sure that we really mean that the point about productivity follows naturally from the idea that researchers have identified a need for transparency. I think many would read "it follows that" as "because of what we just said, you can conclude that". If that is what we're trying to say, then we need to make that logical connection clearer. I think, though, that we're trying to say that another plus of using a computational reproducible approach (and the tools that go with it) is that it makes research more efficient and collaboration easier. So more of an "also" that a "therefore". (See the Issue that I've added, too, about potentially editing this paragraph a bit more to make the meta-discourse, like "it follows", "thus", etc., line up more with what I think is the point we're trying to make.)