Change "to optimize your workflow" to "Before looking for a package to use for a task, consider whether you need one"
I agree that there's room in this rule (and it's helpful) to also talk about how considering your purpose can direct you to the right package, if you do need a package. However, I think that the first (and nice and surprising!) message in this rule is that *you might not need anything!* I want us to make sure we don't bury that, and that we get the reader there very quickly. I think that we can do that here by framing that as one reason they should think about their purpose, and then later (after we talk about whether or not you need a package) develop into how knowing their purpose can help in picking a package if they do need one.

If we keep "workflow", let's think about that wording. Do we mean optimizing the workflow for the final data analysis they'll do, or the workflow of finding the right package? I think a lot of readers for this journal would assume the first, but either way, there could be some confusion so we might want to think if we could clarify with a different word choice.