Take out some redundant words
I think we're saying the same thing if we take out these words. I took out "first", "task and", "then", and "of the task". Also, I changed "steps" to "tasks". We later use "tasks" a lot, so I wanted to make sure that we are being consistent in our terminology so that readers will know we mean the same thing. Also, I think that throughout we're using "task" or "tasks" in the sense of smaller, discrete pieces of work that need to get done to get to the goal. In this sense, I think that "tasks" and "steps" are about the same thing.