Take out a sentence
I've taken out the sentence that starts "That is, some...". I don't think this was adding anything that we hadn't already said, and I think it's nice for us to move quickly from the claim that sometimes an R package might not be appropriate to the reason that there are advantages to coding in base R (which I think is the issue statement of this sentence---the one that starts "There are advantages").