Add code formatting for "for" in "for loop"
I got hung up on "for" in this sentence because my brain at first thought we meant it in its usual sense. One solution to help readers is to use code typeface, so that their brains immediately understand we're using "for" in the context of something for code. Another solution would be to use "for-loop" instead of "for loop". That would immediately set this as a place where we don't mean "for" in the usual sense.