Cut a sentence
I don't think we need this sentence (starts with "Many processes...") and are more streamlined without it. Also, if we keep it, we should re-work the wording. Technically, I think that all R packages (and all R code) include processes that involve data. The next sentence's narrower focus on "data manipulation" is clearer in capturing what we want to focus on in this example.