Move two sentences to the Introduction
Both of these sentences in Rule 2 "Packages are essential..." and "Those who have used R packages" seem to be covering the broader topic of the whole paper, rather than Rule 2 specifically, so I think they could work better in the Introduction. We can smooth out the transition here as we edit.