Hone first paragraph of Rule 2
I've made some edits here to try to hone this paragraph in to the specific suggestion in this rule of looking for leads in the different places. The edits here include (1) moving part of one sentence that was more general (new users can struggle to find and evaluate R packages) up to the Introduction; (2) moving the sentence saying that there's info on R packages lots of places to the start of the paragraph; and (3) adding a transition from that---long-time R users know this and know where to look, but if you're new you might not know these spots, so we can help. For this rule, I think we could convey the idea that people who have used R for a while end up with some good "fishing holes" that you may not know how to find if you're just starting out, and so here's a map to some to help you get started. I think we'll still probably want to do some line edits to make the paragraph flow now that I've moved some pieces around, but I think this order helps lead the reader into our point for the rule.