Broaden "collaborate"
With the content we have for the "collaborate" part, I think that word might be too narrow. We're also including community-level stuff that wouldn't involve a direct collaboration (in the sense often meant for scientific research). I've made a suggested change to encompass that this is more leveraging the R community as a whole.