Clarify that GitHub isn't a repo
GitHub and GitLab allow people to post packages, but they aren't a repository. I think (although I'm not positive---we could look it up) that a repository also plays roles of (1) connecting with all the dependencies you need (so when you download a package, it will also download any packages that package uses in its code, if you don't already have it) and (2) provides not just source code, but also binary code, already built for your operating system (for example, if you download from Windows you'll get a binary program file for Windows, and on a Mac one of Mac's OS). By contrast, the code on GitHub is put up as source code, so when you bring it onto your computer, your computer has to build it into the binary file. This is why you have to have devtools when you get a package from CRAN, and why you have to have RTools (for Windows) or XCode (for Mac) to get devtools, because these have compilers in case some of the code is in something like C.