Move some things around in the CRAN paragraph
For this paragraph, I think that are three main points about CRAN are: (1) it's huge, with loads of packages on loads of different things (love the Joe Rickert quote for this); (2) it's very easy to use in conjunction with base R; and (3) its volunteers perform some checks on incoming packages, and some regular checks on existing packages, to maintain some standards of "not-brokenness".

I think the first point is the strongest to lead with. In this commit, I'm moving that earlier in the paragraph, then moving to the point on easy integration with base R.