Work GitLab into the start of GitHub paragraph
I think that GitHub and GitLab are pretty synonymous in the context of sharing packages. Some of the comments we have for GitLab later (git-based, collaborative cloud, ability for continuous integration and deployment) also apply to GitHub, especially with the GitHub Actions features. I've made some edits here to work in GitLab as we start the paragraph, then funnel down to GitHub specifically, since that is the much more popular of the two for R package, I think.