Add something about having to build GitHub packages from source
Almost always, the only thing posted on GitHub for an R package is the source code, not the binaries (as are posted on CRAN). That's why you get all the messages about building from source when you use `install_github` and why you have to have XCode or Rtools to be able to get `devtools`. This is a disadvantage that's probably worth mentioning. The only exception is when people post their drat repositories on GitHub, as these are designed to help share binaries for R packages. This is much less common, though, then just posting the code.