Move all the `drat` info together
Even though GitHub can be used as the platform for a drat repository, I really think of them more as repositories, and as a very different style of sharing packages that posting the source code as a direct GitHub repo. That being the case, I think it makes sense to move all the drat info into this "other repos" paragraph and remove from the GitHub paragraph.