Move the comment about the GitHub / CRAN debate later
This seemed to be a subtler point and tied to the question about handling package dependencies (could you confirm that I've understood our referencing of these sources correctly? it's @mcelreath2020 and @rickert2018). If that's the case, I think this comment works well later in the paragraph when we talk about that facets of advantages / disadvantages of CRAN vs. GitHub.