Flesh out explanation of GitHub-style platforms
For our target audience, many might not know what GitHub is, so it's worth trying to paint a picture for them. I added one sentence here---my main point is that you can almost look over the developer's shoulder as they're coding the package, and that these platforms are really giving you a window into everything that happens in that repository. For readers who don't have a clear mental picture of what a package is (essentially, a directory of files that follows a few rules for what subdirectories it can have and what goes in them and then includes code for R functions as well as some extras like help documentation and tests), it might be hard to visualize how something like GitHub works. We might need to play around with this a bit more as we continue to edit, but maybe it would also help to define/explain how packages are structured somewhere earlier on, so we can refer back to that here when we explain that GitHub lets you basically chuck that whole directory up online for people to explore and watch as you develop it?