Move information about DESCRIPTION to rule 7
I definitely agree we should talk about package metadata, but this seemed to fit in better with this rule about finding out who wrote the package (plus, that rule was shorter on content compared to others, so moving helped balance a bit). I have move the DESCRIPTION content to that rule and added a bit about how to find this metadata online for packages on CRAN or Bioconductor. I wrote this extra kind of quickly as a place holder, so we might want to smooth it out as we edit.