Draft a paragraph on two types of help resources
I think that, for this rule, it may be helpful to frame the help resources as being in two broad categories---the stuff that is embedded in the package source code, and the stuff that's outside that source code. We've already got this split inherent in the order of our paragraphs, but it might be helpful to be even more explicit about that framework.

I've drafted a paragraph here on that point. I think that we ultimately will want to work this into the first paragraph of this rule, but I kept it separate for the moment so we can think through how to do that.