Add some to the paragraph on helpfiles
With the DOCUMENTATION moved to a later rule, I've added some here to flesh out the info on function helpfiles. Most of what I included was to help make the connections to newer users with the content that was already in the paragraph. Also, this is the only help documentation that is clearly at the levels of functions, rather than the package as a whole, so I tried to be a bit more explicit about that characteristic of this type of help.