Add a paragraph to break between "ships with package" help and other help
I imagine that RDocumentation is pulling together the help information that is included in the package itself, like help files and vignettes, but not scouring the internet for outside resources for the package. Is that right? If so, I think that we could mention it in a paragraph to sum up the types of documentation that come with the package, before we move into the part about outside resources. I've added this paragraph as a first pass at doing that, but we can edit it as needed.