Add a closing paragraph to Rule 5
Package downloads are a great way to show that people are trying out your package, but they don't say anything about whether, once they try it, they like it. For really general packages, more downloads often means a better package, because people will start suggesting that package to friends and teaching it in classes and whatnot. However, I do think that it's helpful if we add this paragraph to caveat the metrics we've just talked about, and then that also transitions us into the next rule.