Flesh out steps in testing package code
I think it might be useful here for us to walk the user through how to get started with a vignette and then with help files. I think a good progression is to try out the examples they give with their own example data (which *should* work!) and then try the same functions with your own data. I've added some text here already, but it might also be useful to flesh out this idea of the example data that comes with a package. This could include how you load it (with the `data` function), how it will often be used in examples in the help file or vignette, and then how you can check it out to see what type of object it is (`class`, for example) and what components it has (`str` can help with that), so you can compare it to your own data to diagnose what the input requirements are for that function.