Some edits at the sentence and wording level for paragraph 1.
Notes on these: (1) moving "computational reproducibility" to end of first sentence (instead of beginning) since it's the new idea we want to emphasize; (2) In second sentence, move idea of time ("of our time" to "no longer") to earlier in the sentence---repeating the theme of new/surfacing/emerging trend; (3) Move two sentences earlier in the paragraph, so everything on the benefits of comp. repro. for itself are together, then the side benefits are at the end; (4) Make some of the language in the sentence about productivity/collaboration more conversational.