Add some more small edits for flow in the Intro
This was in really good shape. I've tweaked some here and there to enhance the flow of the "funnel" in getting us through from the opening to the challenge.

You've got a really nice Introduction here now. When you have a chance, go back through the sections in "Writing for Science" on Introductions (on day 2 of our workshop, I think)---make sure you can solidify in your mind how your very nice intro is doing all the jobs described there, so you can do your drafting and edits to get as nice of a product in future papers, too. Really nice work!