Add an Author Summary
So, in looking through example papers, it looks like this isn't a summary of us (the authors), but rather a summary we write that's a slightly more informal summary of what we're doing in the article---kind of a less formal abstract.

We can keep or cut this. In the first example I looked at for a Ten Simple Rules article, it did not look like they included one, but it probably wouldn't hurt to include one if we're allowed. It's another chance for us to say the "so what" (i.e., why did we write this article, and why should the journal publish it).